% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openeo_start-jobs.R
\name{check_job_status}
\alias{check_job_status}
\title{Helper function OpenEO: check job status}
\usage{
check_job_status(job_id)
}
\arguments{
\item{job_id}{job id}
}
\value{
status string of job
}
\description{
Helper function OpenEO: check job status
}
