% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copernicus-cds.R
\name{copernicus_cds}
\alias{copernicus_cds}
\title{Copernicus Climate Data Store: Single Query}
\usage{
copernicus_cds(
  variable = "2m_temperature",
  dataset_short_name = "reanalysis-era5-single-levels",
  product_type = "reanalysis",
  years = as.character(seq(2010, 2020)),
  area = c(40, 116, 39, 117),
  file_format = "grib",
  export_dir = "."
)
}
\arguments{
\item{variable}{variable to query "2m_temperature"}

\item{dataset_short_name}{"reanalysis-era5-single-levels"}

\item{product_type}{default: "reanalysis"}

\item{years}{character vector of years (default:  as.character(seq(2010, 2020)))}

\item{area}{area coordinates in latitude/longitude (default: c(40, 116, 39, 117))}

\item{file_format}{"grib" or "netcdf"}

\item{export_dir}{default: "."}
}
\value{
path to file with exported data
}
\description{
Run request (may take a while as nothing happens because many requests R
are queued (in addition: maximum request data amount limited to 140000
data points).
To check queue status: "https://cds.climate.copernicus.eu/cdsapp#!/yourrequests?tab=form"
}
