% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copernicus-cds.R
\name{copernicus_cds_parallel}
\alias{copernicus_cds_parallel}
\title{Copernicus Climate Data Store: Multi Query}
\usage{
copernicus_cds_parallel(
  variables = c("2m_temperature", "evaporation", "potential_evaporation",
    "precipitation_type", "runoff", "sub_surface_runoff", "surface_runoff",
    "total_precipitation")
)
}
\arguments{
\item{variables}{default: c("2m_temperature", "evaporation", "potential_evaporation",
"precipitation_type", "runoff", "sub_surface_runoff", "surface_runoff",
"total_precipitation")}
}
\value{
list with paths to downloaded files
}
\description{
Copernicus Climate Data Store: Multi Query
}
