% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rgee_environment.R
\name{create_ad4gd_env}
\alias{create_ad4gd_env}
\title{Create AD4GD Environment for Google Earth Engine Satellite Data}
\usage{
create_ad4gd_env(force = FALSE, debug = FALSE)
}
\arguments{
\item{force}{force installation even if already installed}

\item{debug}{print debug messages (default: FALSE)}
}
\value{
python environment required for R package "rgee"
}
\description{
Create AD4GD Environment for Google Earth Engine Satellite Data
}
