% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_periods_in_year.R
\name{create_periods_in_year}
\alias{create_periods_in_year}
\title{Helper function: create periods in year}
\usage{
create_periods_in_year(year, n_year_splits)
}
\arguments{
\item{year}{year (e.g. 2020)}

\item{n_year_splits}{number of periods to separate the year}
}
\value{
data frame with columns "start" and "end"
}
\description{
Helper function: create periods in year
}
\examples{
create_periods_in_year(2018, 4L)
}
\keyword{internal}
