% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata_gee.R
\name{gee_get_metadata}
\alias{gee_get_metadata}
\title{Google Earth Engine: get metadata for collection}
\usage{
gee_get_metadata(collection)
}
\arguments{
\item{collection}{collection}
}
\value{
tibble with metadata (to do: improve format of "geometry" column)
}
\description{
Google Earth Engine: get metadata for collection
}
