% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata_era5.R
\name{get_metadata_era5}
\alias{get_metadata_era5}
\title{Copernicus: get metadata for ERA5}
\usage{
get_metadata_era5(grib_file)
}
\arguments{
\item{grib_file}{path to GRIB file}
}
\value{
metadata
}
\description{
Copernicus: get metadata for ERA5
}
\seealso{
Code taken from https://gis.stackexchange.com/a/360652
}
