% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openeo_start-jobs.R
\name{get_number_of_active_jobs}
\alias{get_number_of_active_jobs}
\title{Helper function OpenEO: get number of active jobs}
\usage{
get_number_of_active_jobs()
}
\value{
number of active jobs
}
\description{
Helper function OpenEO: get number of active jobs
}
