% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_rds.R
\name{import_rds}
\alias{import_rds}
\title{Helper function: import .rds files from folder}
\usage{
import_rds(
  rds_dir,
  flatten = TRUE,
  cols_unnest = "satellite_data_metadata",
  drop_cols = TRUE,
  run_parallel = TRUE,
  n_cores = parallel::detectCores() - 1L,
  debug = TRUE
)
}
\arguments{
\item{rds_dir}{path to directoy containing .rds files}

\item{flatten}{should imported results be flattened (default: TRUE)}

\item{cols_unnest}{vector with names of columns to unnest (default:
"satellite_data_metadata"), only used if flatten == TRUE}

\item{drop_cols}{should columns not used for unnesting be dropped (refers to
the columns: "satellite_data_metadata", "satellite_data", "satellite_metadata"),
default: TRUE}

\item{run_parallel}{should multiple cores (if available) be used for import?
(default: TRUE)}

\item{n_cores}{number of cores to use if \code{run_parallel = TRUE}.
Default: one less than the detected number of cores.}

\item{debug}{print debug messages? (default: TRUE)}
}
\value{
list of imported .rds file or flattened (if flatten == TRUE) with
tibble with column names satellite_data.xxx and satellite_metadata.xxx
}
\description{
Helper function: import .rds files from folder
}
