% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openeo_download-data.R
\name{openeo_download_results}
\alias{openeo_download_results}
\title{Openeo: download results}
\usage{
openeo_download_results(job_id, tdir = tempdir())
}
\arguments{
\item{job_id}{job id}

\item{tdir}{target directory (default: tempdir())}
}
\value{
list with job metadata and data. names are the title of the job
}
\description{
Openeo: download results
}
\examples{
\dontrun{
jobs_finished <- tibble::as_tibble(openeo::list_jobs()) \%>\%
dplyr::filter(status == "finished") \%>\%
dplyr::arrange(dplyr::desc(updated))

sat_data <- lapply(1:6, function(i) {
kwb.satellite::openeo_download_results(job_id = jobs_finished$id[i])
}
)
}
}
