% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openeo_get-data.R
\name{openeo_get_data}
\alias{openeo_get_data}
\title{OpenEO: get data}
\usage{
openeo_get_data(
  lakes,
  collection_id = "SENTINEL2_L2A",
  date_start = "2017-01-01",
  date_end = "2023-12-31",
  point_on_surface = FALSE,
  spatial_fun = "mean",
  bands = NULL,
  output_format = "CSV",
  col_lakename = "SEE_NAME",
  col_lakeid = "SEE_KZ",
  start_job = FALSE
)
}
\arguments{
\item{lakes}{sf object with lake(s) to get data for}

\item{collection_id}{collection id used by OpenEO (default: "SENTINEL2_L2A")}

\item{date_start}{date start (default:  "2017-01-01")}

\item{date_end}{date end (default: "2023-12-31")}

\item{point_on_surface}{(default: "FALSE)}

\item{spatial_fun}{function for spatial aggregation (default: "mean"). If NULL,
raw data are returned in "netCDF" format}

\item{bands}{(default: NULL). Be aware that naming differs compared to Google
Earth Engine (e.g. B01 -> openEO, B1 -> GEE)}

\item{output_format}{default: JSON, in case spatially unaggregated data (i.e.
spatial_fun == NULL) are returned, it is outmatically set to "netCDF"}

\item{col_lakename}{col_lakename ("GEWNAME", used by Berlin authority for surface
water bodies) use "SEE_NAME" for Brandenburg lakes (default: "SEE_NAME")}

\item{col_lakeid}{("GEWRNEU", used by Berlin authority for surface
water bodies) use "SEE_KZ" for Brandenburg lakes (default: "SEE_KZ")}

\item{start_job}{should job be started (default: FALSE)}
}
\value{
list with job object with job id for processing results data or starting
job afterwards and metadata for selected bands
}
\description{
OpenEO: get data
}
\examples{
\dontrun{
openeo_con <- openeo::connect(host = "https://openeo.dataspace.copernicus.eu")
openeo::login(openeo_con)
test_001002 <- openeo_get_data(lakes = lakes_bb_selected[1:2,])
test_001010 <- lapply(11:13, function(i) openeo_get_data(lakes = lakes_bb_selected[i,]))
}
}
