% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openeo_get_job_metadata.R
\name{openeo_get_job_metadata}
\alias{openeo_get_job_metadata}
\title{OpenEO Helper Function: get job metadata}
\usage{
openeo_get_job_metadata(job_id, debug = TRUE)
}
\arguments{
\item{job_id}{job id}

\item{debug}{print debug messages? (default: TRUE)}
}
\value{
tibble with job metadata (e.g. costs)
}
\description{
OpenEO Helper Function: get job metadata
}
