% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openeo_start-jobs.R
\name{openeo_start_max_jobs}
\alias{openeo_start_max_jobs}
\title{OpenEO: start maximum number of jobs}
\usage{
openeo_start_max_jobs(job_ids, max_jobs = 2, check_interval = 30, debug = TRUE)
}
\arguments{
\item{job_ids}{character vector of job ids}

\item{max_jobs}{maximum number of concurrent jobs (default: 2)}

\item{check_interval}{check intervall in seconds if jobs are already finished}

\item{debug}{print debug messages (default: TRUE)}
}
\value{
starts all jobs respecting the max_jobs limit
}
\description{
OpenEO: start maximum number of jobs
}
