% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_to_geojson.R
\name{sf_to_geojson}
\alias{sf_to_geojson}
\title{sf to GeoJSON for OpenEO}
\usage{
sf_to_geojson(sf)
}
\arguments{
\item{sf}{sf object}
}
\value{
geojson formats that works with OpenEO
}
\description{
sf to GeoJSON for OpenEO
}
