% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_year.R
\name{split_year}
\alias{split_year}
\title{Helper function: split year}
\usage{
split_year(year, num_periods)
}
\arguments{
\item{year}{year (e.g. 2020)}

\item{num_periods}{number of periods to separate the year}
}
\value{
data frame with columns "start" and "end"
}
\description{
Helper function: split year
}
\examples{
split_year(2024, 12)
split_year(2018, 6)
}
