% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata_gee.R
\name{convert_to_list}
\alias{convert_to_list}
\title{Helper function: convert to list}
\usage{
convert_to_list(coords)
}
\arguments{
\item{coords}{column with satellite metadata coordinates (system:footprint`$coordinates)}
}
\value{
list of coordinates
}
\description{
Helper function: convert to list
}
\keyword{internal}
