% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copernicus-cds.R
\name{copernicus_cds_parallel}
\alias{copernicus_cds_parallel}
\title{Copernicus Climate Data Store: Multi Query}
\usage{
copernicus_cds_parallel(
  variables = c("2m_temperature", "evaporation", "potential_evaporation",
    "precipitation_type", "runoff", "sub_surface_runoff", "surface_runoff",
    "total_precipitation"),
  dataset_short_name = "reanalysis-era5-single-levels",
  product_type = "reanalysis",
  years = as.character(seq(2010, 2020)),
  area = c(40, 116, 39, 117),
  file_format = "grib",
  export_dir = "."
)
}
\arguments{
\item{variables}{default: c("2m_temperature", "evaporation", "potential_evaporation",
"precipitation_type", "runoff", "sub_surface_runoff", "surface_runoff",
"total_precipitation")}

\item{dataset_short_name}{"reanalysis-era5-single-levels"}

\item{product_type}{default: "reanalysis"}

\item{years}{character vector of years (default:  as.character(seq(2010, 2020)))}

\item{area}{area coordinates in latitude/longitude (default: c(40, 116, 39, 117))}

\item{file_format}{"grib" or "netcdf"}

\item{export_dir}{default: "."}
}
\value{
list with paths to downloaded files
}
\description{
Runs \code{copernicus_cds} in parallel for all variables
defined in parameter "variables" on all machine cores minus one
}
