% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_periods_in_year.R
\name{create_periods_in_year}
\alias{create_periods_in_year}
\title{Helper function: create periods in year}
\usage{
create_periods_in_year(year, n_periods = 4L)
}
\arguments{
\item{year}{year (e.g. 2020)}

\item{n_periods}{number of periods into which to separate the year.
Default: 4L}
}
\value{
data frame with columns "start" and "end"
}
\description{
Helper function: create periods in year
}
\examples{
create_periods_in_year(2018, n_periods = 4L)
}
\keyword{internal}
