% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_results.R
\name{flatten_results}
\alias{flatten_results}
\title{Flatten Results: simplify satellite data and metadata structure by unnesting
from list to tibble}
\usage{
flatten_results(sat_data_list)
}
\arguments{
\item{sat_data_list}{satellite data list result as retrieved by for example
\code{\link{gee_get_data_for_years_parallel}}}
}
\value{
tibble with column names satellite_data.xxx and satellite_metadata.xxx
in addition also the complex column satellite_data.geometry is simplified and
stored in a satellite_data.geometry_coords
}
\description{
Flatten Results: simplify satellite data and metadata structure by unnesting
from list to tibble
}
