% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google-earth-engine_copernicus_get-data.R
\name{gee_get_data}
\alias{gee_get_data}
\title{Google Earth Engine: get data for years}
\usage{
gee_get_data(
  collection,
  lakes,
  point_on_surface = FALSE,
  spatial_fun = "mean",
  scale = 10,
  via = "getInfo",
  col_lakename = "SEE_NAME",
  debug = TRUE
)
}
\arguments{
\item{collection}{collection satellite collection}

\item{lakes}{lakes sf data frame witch shapes of lakes}

\item{point_on_surface}{use sf::st_point_on_surface() or polygon? (default: FALSE)}

\item{spatial_fun}{spatial aggregation function (default: "mean")}

\item{scale}{scale parameter (default: 10), for details, see
\url{https://developers.google.com/earth-engine/guides/scale}}

\item{via}{via (default: "getInfo"), other options use google cloud storage}

\item{col_lakename}{col_lakename ("GEWNAME", used by Berlin authority for surface
water bodies) use "SEE_NAME" for Brandenburg lakes (default: "SEE_NAME")}

\item{debug}{print debug messages? (default: TRUE)}
}
\value{
tibble
}
\description{
Google Earth Engine: get data for years
}
