% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google-earth-engine_copernicus_get-data.R
\name{gee_get_data}
\alias{gee_get_data}
\title{Google Earth Engine: get data for years}
\usage{
gee_get_data(
  collection,
  lakes,
  centroid = FALSE,
  ee_fun = rgee::ee$Reducer$mean(),
  via = "getInfo",
  col_lakename = "GEWNAME",
  debug = TRUE
)
}
\arguments{
\item{collection}{collection}

\item{lakes}{lakes}

\item{centroid}{centroid}

\item{ee_fun}{ee_fun}

\item{via}{via}

\item{col_lakename}{col_lakename}

\item{debug}{debug}
}
\value{
tibble
}
\description{
Google Earth Engine: get data for years
}
