% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google-earth-engine_copernicus_get-data.R
\name{gee_get_data_for_years}
\alias{gee_get_data_for_years}
\title{Google Earth Engine: get data for years}
\usage{
gee_get_data_for_years(
  years = 2018,
  lakes,
  image_collection = "COPERNICUS/S2_SR_HARMONIZED",
  bands = NULL,
  point_on_surface = FALSE,
  spatial_fun = "mean",
  scale = 10,
  via = "getInfo",
  col_lakename = "SEE_NAME",
  debug = TRUE,
  ee_print = FALSE,
  n_year_splits = NULL
)
}
\arguments{
\item{years}{years vector of years for which satellite data should be downloaded}

\item{lakes}{lakes sf data frame witch shapes of lakes}

\item{image_collection}{image collection (default: "COPERNICUS/S2_SR_HARMONIZED")}

\item{bands}{bands (defualt: NULL), for selection provide in the following
format: as.list(c("QA60", paste0("B", 1:6)))}

\item{point_on_surface}{use sf::st_point_on_surface or polygon? (default: FALSE)}

\item{spatial_fun}{spatial aggregation function (default: "mean")}

\item{scale}{scale parameter (default: 10), for details, see
\url{https://developers.google.com/earth-engine/guides/scale}}

\item{via}{via (default: "getInfo"), other options use google cloud (google drive
or google cloud storage)}

\item{col_lakename}{col_lakename ("GEWNAME", used by Berlin authority for surface
water bodies) use "SEE_NAME" for Brandenburg lakes (default: "SEE_NAME")}

\item{debug}{print debug messages? (default: TRUE)}

\item{ee_print}{show debug messages for "ee" (default: FALSE)}

\item{n_year_splits}{number of year splits per request. Required in case request
uses too much images > 400-500 per year (default: NULL, determined automatically within
function. In case it should be overwritten by the user provide a meaningful integer number)}
}
\value{
list with data and metadata, each of them tibbles
}
\description{
Google Earth Engine: get data for years
}
