% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google-earth-engine_copernicus_get-data.R
\name{gee_get_data_for_years}
\alias{gee_get_data_for_years}
\title{Google Earth Engine: get data for years}
\usage{
gee_get_data_for_years(
  years = 2018,
  lakes,
  bands = as.list(c("QA60", paste0("B", 1:6))),
  centroid = FALSE,
  ee_fun = rgee::ee$Reducer$mean(),
  via = "getInfo",
  col_lakename = "GEWNAME",
  debug = TRUE,
  ee_print = FALSE
)
}
\arguments{
\item{years}{years}

\item{lakes}{lakes}

\item{bands}{bands}

\item{centroid}{centroid}

\item{ee_fun}{ee_fun}

\item{via}{via}

\item{col_lakename}{col_lakename (default: "GEWNAME")}

\item{debug}{show debug messages (default: TRUE)}

\item{ee_print}{show debug messages for "ee" (default: FALSE)}
}
\value{
tibble
}
\description{
Google Earth Engine: get data for years
}
