% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/google-earth-engine_copernicus_get-data.R
\name{gee_get_data_for_years}
\alias{gee_get_data_for_years}
\title{Google Earth Engine: get data for years}
\usage{
gee_get_data_for_years(
  years = 2018,
  lakes,
  image_collection = "COPERNICUS/S2_SR_HARMONIZED",
  bands = as.list(c("QA60", paste0("B", 1:6))),
  centroid = FALSE,
  ee_fun = rgee::ee$Reducer$mean(),
  scale = 10,
  via = "getInfo",
  col_lakename = "GEWNAME",
  debug = TRUE,
  ee_print = FALSE
)
}
\arguments{
\item{years}{years vector of years for which satellite data should be downloaded}

\item{lakes}{lakes sf data frame witch shapes of lakes}

\item{image_collection}{image collection (default: "COPERNICUS/S2_SR_HARMONIZED")}

\item{bands}{bands}

\item{centroid}{use centroid or polygon? (default: FALSE)}

\item{ee_fun}{spatial aggregation function (default: rgee::ee$Reducer$mean())}

\item{scale}{scale parameter (default: 10), for details, see
\url{https://developers.google.com/earth-engine/guides/scale}}

\item{via}{via (default: "getInfo"), other options use google cloud (google drive
or google cloud storage)}

\item{col_lakename}{col_lakename (default: "GEWNAME")}

\item{debug}{show debug messages (default: TRUE)}

\item{ee_print}{show debug messages for "ee" (default: FALSE)}
}
\value{
list with data and metadata, each of them tibbles
}
\description{
Google Earth Engine: get data for years
}
