% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/google-earth-engine_copernicus_get-data_parallel.R
\name{gee_get_data_for_years_parallel}
\alias{gee_get_data_for_years_parallel}
\title{Google Earth Engine: get data for years in parallel}
\usage{
gee_get_data_for_years_parallel(
  years = 2018,
  lakes,
  image_collection = "COPERNICUS/S2_SR_HARMONIZED",
  bands = as.list(c("QA60", paste0("B", 1:6))),
  centroid = FALSE,
  ee_fun = rgee::ee$Reducer$mean(),
  scale = 10,
  via = "getInfo",
  col_lakename = "GEWNAME",
  set_lakenames_as_list_indices = TRUE,
  debug = TRUE,
  debug_dir = tempdir(),
  ee_print = FALSE,
  ncores = parallel::detectCores() - 1
)
}
\arguments{
\item{years}{years vector of years for which satellite data should be downloaded}

\item{lakes}{lakes sf data frame witch shapes of lakes}

\item{image_collection}{image collection (default: "COPERNICUS/S2_SR_HARMONIZED")}

\item{bands}{bands}

\item{centroid}{use centroid or polygon? (default: FALSE)}

\item{ee_fun}{spatial aggregation function (default: rgee::ee$Reducer$mean())}

\item{scale}{scale parameter (default: 10), for details, see
\url{https://developers.google.com/earth-engine/guides/scale}}

\item{via}{via (default: "getInfo"), other options use google cloud (google drive
or google cloud storage)}

\item{col_lakename}{col_lakename (default: "GEWNAME")}

\item{set_lakenames_as_list_indices}{should lake names of "col_lakename" be used
for naming result list? (default: TRUE)}

\item{debug}{show debug messages (default: TRUE)}

\item{debug_dir}{directory where to save (default: tempdir())}

\item{ee_print}{show debug messages for "ee" (default: FALSE)}

\item{ncores}{number of cores for parallel processinfg (default:
parallel::detectCores() - 1)}
}
\value{
list with data and metadata, each of them tibbles
}
\description{
Google Earth Engine: get data for years in parallel
}
