% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata_gee.R
\name{gee_get_metadata}
\alias{gee_get_metadata}
\title{Google Earth Engine: get metadata for collection}
\usage{
gee_get_metadata(collection)
}
\arguments{
\item{collection}{collection}
}
\value{
tibble with metadata. In addition also the complex column
"geometry_meta_org" is simplified and stored in the new column "geometry_meta_cleaned"
}
\description{
Google Earth Engine: get metadata for collection
}
