% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_rds.R
\name{import_rds}
\alias{import_rds}
\title{Helper function: import .rds files from folder}
\usage{
import_rds(rds_dir)
}
\arguments{
\item{rds_dir}{path to directoy containing .rds files}
}
\value{
list with imported
}
\description{
Helper function: import .rds files from folder
}
