% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openeo_get-data.R
\name{openeo_get_bands_meta}
\alias{openeo_get_bands_meta}
\title{OpenEO Helper function: get bands metadata}
\usage{
openeo_get_bands_meta(collection_id)
}
\arguments{
\item{collection_id}{collection id}
}
\value{
tibble with bands metadata
}
\description{
OpenEO Helper function: get bands metadata
}
\examples{
\dontrun{
openeo_get_bands_meta(collection_id = "SENTINEL2_L2A")
}

}
