% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openeo_get-data.R
\name{openeo_get_bands_meta}
\alias{openeo_get_bands_meta}
\title{OpenEO Helper function: get bands metadata}
\usage{
openeo_get_bands_meta(collection_id, simplify = TRUE)
}
\arguments{
\item{collection_id}{collection id}

\item{simplify}{if TRUE aliases column will be collapsed so that multiple aliases
were joined with '/' and resulting tibble contains no sublists anymore
(default: TRUE)}
}
\value{
tibble with bands metadata
}
\description{
OpenEO Helper function: get bands metadata
}
\examples{
\dontrun{
openeo_get_bands_meta(collection_id = "SENTINEL2_L2A")
}

}
